+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * mysql: updated and patched to 8.3.0, fixed CJException should be wrapped as SQLException
   > make sure use "core.framework.mysql:mysql-connector-j:8.3.0-r2"
 * thread: replace all synchronized with ReentrantLock/Condition, including RateControl/ShutdownHandling/Test
+* http: convert some http error as warning
+  > undertow "UT000133: Request did not contain an Upgrade header, upgrade is not permitted"
+  > "response was sent, discard the current http transaction"
 
 ### 9.0.8 (1/29/2024 - 3/7/2024)